chore: refactor marker and filter structure in map
