+++ frontend/src/routes/_protected/map/index.tsx
@@ -6,18 +6,19 @@ import {
 import MapButtons from '@/components/map/MapButtons'
 import { Tree, TreeCluster } from '@green-ecolution/backend-client'
 import { useSuspenseQuery } from '@tanstack/react-query'
-import { WithTreesAndClusters } from '@/components/map/TreeMarker'
 import { treeClusterQuery, treeQuery } from '@/api/queries'
 import { useRef, useState } from 'react'
 import Dialog from '@/components/general/filter/Dialog'
 import StatusFieldset from '@/components/general/filter/fieldsets/StatusFieldset'
 import { getWateringStatusDetails } from '@/hooks/useDetailsForWateringStatus'
 import useFilter from '@/hooks/useFilter'
-import FilterProvider from '@/context/FilterContext'
+import FilterProvider, { Filters } from '@/context/FilterContext'
 import { z } from 'zod'
 import ClusterFieldset from '@/components/general/filter/fieldsets/ClusterFieldset'
 import PlantingYearFieldset from '@/components/general/filter/fieldsets/PlantingYearFieldset'
 import useMapInteractions from '@/hooks/useMapInteractions'
+import { WithTreesAndClusters } from '@/components/map/marker/WithAllClusterAndTrees'
+import WithFilterdTrees from '@/components/map/marker/WithFilterdTrees'
 
 const mapFilterSchema = z.object({
   status: z.array(z.string()).optional(),
@@ -28,6 +29,41 @@ const mapFilterSchema = z.object({
   cluster: z.number().optional(),
 })
 
+const hasActiveFilter = (search: {
+  status: string[]
+  hasCluster: true | undefined
+  plantingYears: number[]
+  tree: number | undefined
+  cluster: number | undefined
+}) => {
+  return search.status.length > 0 ||
+    search.hasCluster !== undefined ||
+    search.plantingYears.length > 0
+    ? true
+    : false
+}
+
+const filterData = (trees: Tree[], filters: Filters) => {
+  return trees.filter((tree) => {
+    const statusFilter =
+      filters.statusTags.length === 0 ||
+      filters.statusTags.includes(
+        getWateringStatusDetails(tree.wateringStatus).label
+      )
+
+    const hasCluster =
+      filters.hasCluster === undefined ||
+      (filters.hasCluster && tree.treeClusterId) ||
+      (!filters.hasCluster && !tree.treeClusterId)
+
+    const plantingYearsFilter =
+      filters.plantingYears.length === 0 ||
+      filters.plantingYears.includes(tree.plantingYear)
+
+    return statusFilter && hasCluster && plantingYearsFilter
+  })
+}
+
 function MapView() {
   const navigate = useNavigate({ from: '/map' })
   const search = useLoaderData({ from: '/_protected/map/' })
@@ -36,6 +72,10 @@ function MapView() {
   const { data: cluster } = useSuspenseQuery(treeClusterQuery())
   const { data: trees } = useSuspenseQuery(treeQuery())
   const { filters } = useFilter()
+  const [activeFilter, setActiveFilter] = useState(hasActiveFilter(search))
+  const [filteredData, setFilteredData] = useState<Tree[]>(
+    filterData(trees.data, filters)
+  )
 
   const handleTreeClick = (tree: Tree) => {
     navigate({ to: `/tree/$treeId`, params: { treeId: tree.id.toString() } })
@@ -48,41 +88,8 @@ function MapView() {
     })
   }
 
-  const hasActiveFilter = () => {
-    return search.status.length > 0 ||
-      search.hasCluster !== undefined ||
-      search.plantingYears.length > 0
-      ? true
-      : false
-  }
-
-  const [activeFilter, setActiveFilter] = useState(hasActiveFilter())
-
-  const filterData = () => {
-    return trees.data.filter((tree) => {
-      const statusFilter =
-        filters.statusTags.length === 0 ||
-        filters.statusTags.includes(
-          getWateringStatusDetails(tree.wateringStatus).label
-        )
-
-      const hasCluster =
-        filters.hasCluster === undefined ||
-        (filters.hasCluster && tree.treeClusterId) ||
-        (!filters.hasCluster && !tree.treeClusterId)
-
-      const plantingYearsFilter =
-        filters.plantingYears.length === 0 ||
-        filters.plantingYears.includes(tree.plantingYear)
-
-      return statusFilter && hasCluster && plantingYearsFilter
-    })
-  }
-
-  const [filteredData, setFilteredData] = useState<Tree[]>(filterData())
-
   const handleFilter = () => {
-    const data = filterData()
+    const data = filterData(trees.data, filters)
     if (
       filters.statusTags.length > 0 ||
       filters.hasCluster !== undefined ||
@@ -122,12 +129,21 @@ function MapView() {
         </Dialog>
       </div>
       <MapButtons />
-      <WithTreesAndClusters
-        onClickTree={handleTreeClick}
-        onClickCluster={handleClusterClick}
-        hasHighlightedTree={search.tree}
-        hasHighlightedCluster={search.cluster}
-      />
+      {activeFilter ? (
+        <WithFilterdTrees
+          onClick={handleTreeClick}
+          selectedTrees={search.tree ? [search.tree] : []}
+          hasHighlightedTree={search.tree}
+          filterdTrees={filteredData}
+        />
+      ) : (
+        <WithTreesAndClusters
+          onClickTree={handleTreeClick}
+          onClickCluster={handleClusterClick}
+          hasHighlightedTree={search.tree}
+          hasHighlightedCluster={search.cluster}
+        />
+      )}
     </>
   )
 }
